move utility function from tasks into a separate subscription utils file
